solver: Move _SNESContext to solving_utils
